moved cudaFunctionAnnotation
